+++ LindaEngine/Runtime/Core/src/ApplicationRuntime.cpp
@@ -71,7 +71,7 @@ void ApplicationRuntime::Tick()
             LThread::Sleep(1);
             continue;
         }
-        float deltaTime = currentInterval * 0.001;
+        double deltaTime = currentInterval * 0.001;
         currentInterval = 0.0;
 
         SceneManager::Tick();
@@ -84,7 +84,7 @@ void ApplicationRuntime::Tick()
         BehaviorSystem::DoOnCollisionEvent();
         _window->Tick();
         BehaviorSystem::DoOnMouseEvent();
-        ComponentSystem::Tick(deltaTime);
+        ComponentSystem::Tick((float)deltaTime);
         BehaviorSystem::DoUpdate();
         BehaviorSystem::DoLateUpdate();
